pyproject.toml
[project]
name = "skincare-analysis"
version = "0.1.0"
description = "Skincare analysis for Gen Z on TikTok data"
authors = [
    {name = "Hannah Schlaucher",email = "hanschlau@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "torch[macos] (>=2.6.0,<3.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "notebook (>=7.4.1,<8.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "ipywidgets (>=8.1.6,<9.0.0)",
    "numpy (==1.26.4)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "bertopic (>=0.17.0,<0.18.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "stopwordsiso (>=0.6.1,<0.7.0)",
    "emoji (>=2.14.1,<3.0.0)",
    "rake-nltk (>=1.0.6,<2.0.0)",
    "langdetect (>=1.0.9,<2.0.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "sentencepiece (>=0.2.0,<0.3.0)",
    "google-cloud (>=0.34.0,<0.35.0)",
    "google-cloud-bigquery (>=3.31.0,<4.0.0)",
    "gcloud (>=0.18.3,<0.19.0)",
    "db-dtypes (>=1.4.2,<2.0.0)",
    "streamlit (>=1.45.1,<2.0.0)",
    "watchdog (>=6.0.0,<7.0.0)",
    "circlify (>=0.15.0,<0.16.0)",
    "spacy (>=3.8.7,<4.0.0)",
    "streamlit-option-menu (>=0.4.0,<0.5.0)",
    "gensim (>=4.3.3,<5.0.0)",
    "burst-detection (>=0.1.3,<0.2.0)",
    "spacy (>=3.8.7,<4.0.0)",
    "vadersentiment (>=3.3.2,<4.0.0)",
    "lightgbm (>=4.6.0,<5.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "xgboost (>=3.0.2,<4.0.0)",
    "shap (>=0.48.0,<0.49.0)",
    "wordcloud (>=1.9.4,<2.0.0)",
    "tensorflow-io-gcs-filesystem (==0.36.0)",
]
[tool.poetry]
package-mode = false
[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

